espace de choix courrier

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nom_complet');
             $table->string('email')->unique();
-            $table->unsignedInteger('matricule')->unique();
+            $table->unsignedBigInteger('matricule')->unique();
             $table->enum('role', ['admin', 'user'])->default('user');
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
